v1.40.43 Released!
Request interval now can be set less than 0.5 seconds
Fixed bug when timers wasn't stopped on group stop
Fixed crash when open Debug dialog
Fixed incorrect fee value in Bitfinex
Fixed filter open orders in Bitfinex
Fixed tray icon in Linux
Fixed bug caused long delay after network down
Fixed on top mode for dock widgets
Fixed bug when password was incorrectly detected as weak
Currency pairs synchronized in all exchanges

+++ src/historymodel.h
@@ -65,7 +65,7 @@ public:
 private:
     int dateWidth = 0;
     int typeWidth = 0;
-    quint32 lastDate;
+    qint64 lastDate;
     int columnsCount;
     QStringList headerLabels;
     QStringList typesLabels;
